renamed model config to retinanet_config
